Item summary model
- Added an item summary model to seperate code.
